Services code are refactored
