+++ components/create-employee-modal.tsx
@@ -13,7 +13,7 @@ import {
 } from '@/utils/constants/common-constants';
 import './dropdown-select.css';
 import { useSession } from 'next-auth/react';
-import { ApiService } from '@/services/api-services';
+import { UserService } from '@/services/user-services';
 import { toast } from 'react-toastify';
 
 if (Modal.defaultStyles.overlay) {
@@ -40,7 +40,7 @@ const CreateEmployeeModal = ({
     const getData = async () => {
       //@ts-ignore
       const token = data?.user?.access_token;
-      const Services = new ApiService();
+      const Services = new UserService();
       if (token) {
         const resp = await Services.getManagerList(token);
         const data = resp?.data?.Data?.Data?.map((item: ManagerType) => {
@@ -115,7 +115,7 @@ const CreateEmployeeModal = ({
         //@ts-ignore
         const token = data?.user?.access_token;
 
-        const UserServices = new ApiService();
+        const UserServices = new UserService();
         const resp = await UserServices.createUser(formData, token);
 
         console.log(resp);
