+++ components/leads-list/index.tsx
@@ -11,6 +11,7 @@ import FilterLeadsButton from '../filter-leads-button';
 import CreateLeadsButton from '../create-leads-button';
 import { LeadsContext } from '@/context/leads-context';
 import Loader from '../loader';
+import { UserService } from '@/services/user-services';
 
 function LeadsList() {
   const [searchValue, setSearchValue] = useState<string>('');
@@ -44,7 +45,8 @@ function LeadsList() {
   useEffect(() => {
     if (token) {
       const LeadServices = new LeadService();
-      LeadServices.getExecutivesData(setExecutivesOption, token, setIsLoading);
+      const UserServices = new UserService();
+      UserServices.getExecutivesData(setExecutivesOption, token, setIsLoading);
       LeadServices.getCreatedByData(setCreatedByOptions, token);
       LeadServices.getLeadsData(setLeadsData, token, setIsLoading);
     }
