+++ services/auth-service.ts
@@ -12,6 +12,7 @@ export class AuthService {
     this.client = new HttpClient(baseUrl);
   }
 
+  //* Service for registration
   public signup = async (data: any) => {
     const resp: Response<any> = await this.client.request({
       url: API_PATHS.Signup,
@@ -22,6 +23,7 @@ export class AuthService {
     return resp.data;
   };
 
+  //* Service for login
   public login = async (data: any) => {
     const resp: Response<any> = await this.client.request({
       url: 'auth/login',
@@ -32,6 +34,7 @@ export class AuthService {
     return resp.data;
   };
 
+  //* Service for logout
   public logout = async (token: string): Promise<any> => {
     console.log(token);
     const config: AxiosRequestConfig = {};
@@ -50,6 +53,7 @@ export class AuthService {
     return resp;
   };
 
+  //* Service for forget password
   public forgetPassword = async (data: any) => {
     const resp: Response<any> = await this.client.request({
       url: '/user/forget-password',
@@ -60,6 +64,7 @@ export class AuthService {
     return resp;
   };
 
+  //* Service for reset password
   public resetPassword = async (token: string, company_id: number, data: any) => {
     const resp: Response<any> = await this.client.request({
       url: `/user/reset-password?token=${token}&company_id=${company_id}`,
