+++ services/lead-services.ts
@@ -9,45 +9,7 @@ export class LeadService {
     this.client = new HttpClient(baseUrl);
   }
 
-  public getExecutives = async (token: string): Promise<any> => {
-    const config: AxiosRequestConfig = {};
-
-    if (token) {
-      config.headers = { Authorization: `Bearer ${token}` };
-    }
-    const resp = await this.client.request({
-      url: API_PATHS.GetExecutives,
-      method: API_METHODS.GET,
-      ...config,
-    });
-
-    return resp;
-  };
-
-  public getExecutivesData = async (
-    setExecutivesOption: any,
-    token: string,
-    setIsLoading: (item: boolean) => void
-  ) => {
-    try {
-      const response = await this.getExecutives(token);
-      const executivesOption = this.createSelectData(response.data.Data.Data);
-      setExecutivesOption(executivesOption);
-      setIsLoading(false);
-    } catch (error) {
-      console.error('Error fetching executives:', error);
-    }
-  };
-
-  public createSelectData = (items: any): any => {
-    const selectOptions: any = [];
-    items.map((item: any) => {
-      const newItem = { ...item, value: item.name, label: item.name };
-      selectOptions.push(newItem);
-    });
-    return selectOptions;
-  };
-
+  //* Service to get all lead data
   public getLeads = async (token: string): Promise<any> => {
     const config: AxiosRequestConfig = {};
 
@@ -64,6 +26,7 @@ export class LeadService {
     return resp;
   };
 
+  //* Service to get leads data
   public getLeadsData = async (
     setLeadsData: any,
     token: string,
@@ -81,6 +44,7 @@ export class LeadService {
     }
   };
 
+  //* Service to upload lead images
   public UploadLeadImage = async (data: any, token: string): Promise<any> => {
     const config: AxiosRequestConfig = {};
 
@@ -98,23 +62,7 @@ export class LeadService {
     return resp;
   };
 
-  public createReminder = async (data: any, token: string): Promise<any> => {
-    const config: AxiosRequestConfig = {};
-
-    if (token) {
-      config.headers = { Authorization: `Bearer ${token}` };
-    }
-
-    const resp = await this.client.request({
-      url: API_PATHS.CreateReminder,
-      method: API_METHODS.POST,
-      ...config,
-      data,
-    });
-
-    return resp;
-  };
-
+  //* Service to create lead
   public createLead = async (data: any, token: string): Promise<any> => {
     const config: AxiosRequestConfig = {};
 
@@ -132,6 +80,7 @@ export class LeadService {
     return resp;
   };
 
+  //* Service to delete lead
   public deleteLead = async (lead_id: number, token: string): Promise<any> => {
     const config: AxiosRequestConfig = {};
 
@@ -148,33 +97,29 @@ export class LeadService {
     return resp;
   };
 
+  //* Service to get data of whom who created the lead
   public getCreatedByData = async (setCreatedByOptions: any, token: string) => {
-  try {
-    const response = await this.getLeads(token);
-    const data = response.data.Data.Data;
-    const uniqueCreatedByValues: any[] = [];
-    const encounteredValues = new Set<string>();
+    try {
+      const response = await this.getLeads(token);
+      const data = response.data.Data.Data;
+      const uniqueCreatedByValues: any[] = [];
+      const encounteredValues = new Set<string>();
 
-    data.map((item: any) => {
-      const createdBy = item.created_by;
-      if (createdBy && !encounteredValues.has(createdBy)) {
-        encounteredValues.add(createdBy);
-        const newItem = { ...item, value: item.created_by, label: item.created_by };
-        uniqueCreatedByValues.push(newItem);
-      }
-    });
-    setCreatedByOptions(uniqueCreatedByValues);
-  } catch (error) {
-    console.error('Error fetching leads:', error);
-  }
-};
+      data.map((item: any) => {
+        const createdBy = item.created_by;
+        if (createdBy && !encounteredValues.has(createdBy)) {
+          encounteredValues.add(createdBy);
+          const newItem = { ...item, value: item.created_by, label: item.created_by };
+          uniqueCreatedByValues.push(newItem);
+        }
+      });
+      setCreatedByOptions(uniqueCreatedByValues);
+    } catch (error) {
+      console.error('Error fetching leads:', error);
+    }
+  };
 
-
-
-
-
-  //! Get single Lead data
-
+  //* Service to get single lead data
   public getUserLead = async (user_id: number, token: string): Promise<any> => {
     const config: AxiosRequestConfig = {};
 
@@ -191,8 +136,7 @@ export class LeadService {
     return resp;
   };
 
-  //! Update Lead data
-
+  //* Service to update lead data
   public updateLead = async (lead_id: number, data: any, token: string): Promise<any> => {
     const config: AxiosRequestConfig = {};
 
@@ -210,61 +154,10 @@ export class LeadService {
     return resp;
   };
 
-  public getAllRemindersData = async (token: string, setRemainders: any) => {
-    const res = await this.getAllReminder(token);
-    console.log(res);
-    setRemainders(res?.data?.Data?.Data);
-  };
-
-  public getAllReminder = async (token: string): Promise<any> => {
+  //* Service to filter leads data
+  public FilteredLeadsData = async (data: any, token: string): Promise<any> => {
     const config: AxiosRequestConfig = {};
 
-    if (token) {
-      config.headers = { Authorization: `Bearer ${token}` };
-    }
-
-    const resp = await this.client.request({
-      url: API_PATHS.GetAllReminder,
-      method: API_METHODS.GET,
-      ...config,
-    });
-
-    return resp;
-  };
-
-  public deleteReminder = async (reminder_id: number, token: string): Promise<any> => {
-    const config: AxiosRequestConfig = {};
-
-    if (token) {
-      config.headers = { Authorization: `Bearer ${token}` };
-    }
-
-    const resp = await this.client.request({
-      url: `${API_PATHS.DeleteReminder}?reminder_id=${reminder_id}`,
-      method: API_METHODS.DELETE,
-      ...config,
-    });
-
-    return resp;
-  };
-
-  public updateRemainder = async (reminder_id: number, data:any, token: string): Promise<any> => {
-    const config: AxiosRequestConfig = {}
-    if (token) {
-      config.headers = { Authorization: `Bearer ${token}` };
-    }
-    const resp = await this.client.request({
-      url: `${API_PATHS.UpdateRemainder}?reminder_id=${reminder_id}`,
-      method: API_METHODS.PATCH,
-      ...config,
-      data,
-    });
-    return resp;
-  };
-
-  public FilteredLeadsData = async ( data: any, token: string): Promise<any> => {
-    const config: AxiosRequestConfig = {};
-    
     if (token) {
       config.headers = { Authorization: `Bearer ${token}` };
     }
@@ -277,10 +170,16 @@ export class LeadService {
     return resp;
   };
 
-  public getFilteredLeadsData = async ( setLeadsData: any,setIsLoading: (item: boolean) => void, data:any, token:string) : Promise<any>=>{
+  //* Service to get filtered leads
+  public getFilteredLeadsData = async (
+    setLeadsData: any,
+    setIsLoading: (item: boolean) => void,
+    data: any,
+    token: string
+  ): Promise<any> => {
     try {
       setIsLoading(true);
-      const response = await this.FilteredLeadsData(data,token);
+      const response = await this.FilteredLeadsData(data, token);
       const filteredLeads = response.data.Data.Data;
       console.log(filteredLeads);
       setLeadsData(filteredLeads);
@@ -290,5 +189,20 @@ export class LeadService {
     }
   };
 
+  //* Service to get latest leads
+  public latestLeads = async (token: string): Promise<any> => {
+    const config: AxiosRequestConfig = {};
 
+    if (token) {
+      config.headers = { Authorization: `Bearer ${token}` };
+    }
+
+    const resp = await this.client.request({
+      url: API_PATHS.LatestLeads,
+      method: API_METHODS.GET,
+      ...config,
+    });
+
+    return resp;
+  };
 }
