+++ services/user-services.ts
@@ -2,13 +2,56 @@ import { API_METHODS, API_PATHS } from '@/utils/constants/common-constants';
 import { AxiosRequestConfig } from 'axios';
 import { HttpClient } from './axios-base-query';
 
-export class ApiService {
+export class UserService {
   client;
 
   constructor(baseUrl?: string) {
     this.client = new HttpClient(baseUrl);
   }
 
+  //* Service to get executives
+  public getExecutives = async (token: string): Promise<any> => {
+    const config: AxiosRequestConfig = {};
+
+    if (token) {
+      config.headers = { Authorization: `Bearer ${token}` };
+    }
+    const resp = await this.client.request({
+      url: API_PATHS.GetExecutives,
+      method: API_METHODS.GET,
+      ...config,
+    });
+
+    return resp;
+  };
+
+  //* Service to create select data
+  public createSelectData = (items: any): any => {
+    const selectOptions: any = [];
+    items.map((item: any) => {
+      const newItem = { ...item, value: item.name, label: item.name };
+      selectOptions.push(newItem);
+    });
+    return selectOptions;
+  };
+
+  //* Service to get executives data
+  public getExecutivesData = async (
+    setExecutivesOption: any,
+    token: string,
+    setIsLoading: (item: boolean) => void
+  ) => {
+    try {
+      const response = await this.getExecutives(token);
+      const executivesOption = this.createSelectData(response.data.Data.Data);
+      setExecutivesOption(executivesOption);
+      setIsLoading(false);
+    } catch (error) {
+      console.error('Error fetching executives:', error);
+    }
+  };
+
+  //* Service to create new user
   public createUser = async (data: any, token: string): Promise<any> => {
     const config: AxiosRequestConfig = {};
 
@@ -26,6 +69,7 @@ export class ApiService {
     return resp;
   };
 
+  //* Service to get manager list
   public getManagerList = async (token: string): Promise<any> => {
     const config: AxiosRequestConfig = {};
 
@@ -43,71 +87,7 @@ export class ApiService {
     return resp;
   };
 
-  public createLead = async (data: any, token: string): Promise<any> => {
-    const config: AxiosRequestConfig = {};
-
-    if (token) {
-      config.headers = { Authorization: `Bearer ${token}` };
-    }
-
-    const resp = await this.client.request({
-      url: API_PATHS.CreateLead,
-      method: API_METHODS.POST,
-      ...config,
-      data,
-    });
-
-    return resp;
-  };
-
-  public dashboardInfo = async (token: string): Promise<any> => {
-    const config: AxiosRequestConfig = {};
-
-    if (token) {
-      config.headers = { Authorization: `Bearer ${token}` };
-    }
-
-    const resp = await this.client.request({
-      url: API_PATHS.DashboardInfo,
-      method: API_METHODS.GET,
-      ...config,
-    });
-
-    return resp;
-  };
-
-  public latestLeads = async (token: string): Promise<any> => {
-    const config: AxiosRequestConfig = {};
-
-    if (token) {
-      config.headers = { Authorization: `Bearer ${token}` };
-    }
-
-    const resp = await this.client.request({
-      url: API_PATHS.LatestLeads,
-      method: API_METHODS.GET,
-      ...config,
-    });
-
-    return resp;
-  };
-
-  public leaderboard = async (token: string): Promise<any> => {
-    const config: AxiosRequestConfig = {};
-
-    if (token) {
-      config.headers = { Authorization: `Bearer ${token}` };
-    }
-
-    const resp = await this.client.request({
-      url: API_PATHS.Leaderboard,
-      method: API_METHODS.GET,
-      ...config,
-    });
-
-    return resp;
-  };
-
+  //* Service to get user info
   public getUserInfo = async (token: string): Promise<any> => {
     const config: AxiosRequestConfig = {};
 
@@ -125,6 +105,7 @@ export class ApiService {
     return resp;
   };
 
+  //* Service to get employee list
   public EmployeeListInfo = async (token: string): Promise<any> => {
     const config: AxiosRequestConfig = {};
 
@@ -141,6 +122,7 @@ export class ApiService {
     return resp;
   };
 
+  //* Service to delete a user
   public deleteUser = async (id: number, token: string): Promise<any> => {
     const config: AxiosRequestConfig = {};
 
@@ -157,6 +139,7 @@ export class ApiService {
     return resp;
   };
 
+  //* Service to update employee info
   public updateEmployeeInfo = async (
     user_id: number,
     data: any,
